rename alert_list file
